fix(catalog): use 'double precision' so migrations run on Postgres, not just H2
V1 and V6 used the H2-only 'double' type, which real Postgres rejects. 'double precision' is valid on both dialects, so H2 slice tests keep passing while the schema now also applies on Postgres.

+++ catalog-service/src/main/resources/db/migration/V6__product_detail_dimensions.sql
@@ -1,3 +1,3 @@
-alter table product_detail add column length_cm double;
-alter table product_detail add column width_cm double;
-alter table product_detail add column height_cm double;
+alter table product_detail add column length_cm double precision;
+alter table product_detail add column width_cm double precision;
+alter table product_detail add column height_cm double precision;
